chore: remove type cast

+++ packages/adapters/solana/src/providers/AuthProvider.ts
@@ -170,7 +170,7 @@ export class AuthProvider extends ProviderEventEmitter implements SolanaProvider
 
   // -- Private ------------------------------------------- //
   private serializeTransaction(transaction: AnyTransaction) {
-    return base58.encode(transaction.serialize({ verifySignatures: false }) as Uint8Array)
+    return base58.encode(transaction.serialize({ verifySignatures: false }))
   }
 }
